feat(parser): implement parse_cast_type function and organize module structure
