+++ src/parser/rules/mod.rs
@@ -1,1 +1,2 @@
 pub mod expressions;
+pub mod types;
